authorize
- added option to save paymentoptionresponse
- call retreive auth event
- dialog for extra fees
- enabled fun for extra fees in amount calculator
- reterieve auth api added and integrated

+++ checkout/src/main/java/company/tap/checkout/internal/api/models/SupportedCurrencies.kt
@@ -34,5 +34,5 @@ data class SupportedCurrencies(
     var rate: Double? = null,
     @SerializedName("amount")
     @Expose
-    var amount: Double,
+    var amount: BigDecimal,
     )
